Added rate return to the rate posting

+++ Functions/Rate.ts
@@ -73,8 +73,26 @@ function rate(req: any, res: any, localPath: string) {
 
 	Rates.push(new Rate(r, ip, req.body.pc));
 
+	// Get average rate
+	if (!fs.existsSync(localPath + `rates/${month}/${day}.json`)) {
+		res.status(404).json({ error: 1, msg: 'Rates not found' });
+		return;
+	}
+	const rates = JSON.parse(fs.readFileSync(localPath + `rates/${month}/${day}.json`).toString());
+
+	let average = 0;
+
+	rates.forEach((rate: { rate: number }) => {
+		average += rate.rate;
+	});
+
+	average += r;
+
+	average = parseFloat((average / rates.length + 1).toFixed(1));
+
+	res.status(200).json({ error: 0, msg: 'Success', rate: average });
+
 	fs.writeFileSync(localPath + `rates/${month}/${day}.json`, JSON.stringify(Rates));
-	res.status(200).json({ error: 0, msg: 'Success' });
 }
 
 export default rate;
